requirements.txt
streamlit
youtube-transcript-api
transformers
google-generativeai
python-dotenv
fpdf2
